+++ src/main/kotlin/com/coder/gateway/models/LoginModel.kt
@@ -1,0 +1,3 @@
+package com.coder.gateway.models
+
+internal data class LoginModel(var host: String = "localhost", var port: Int = 7080, var email: String = "example@email.com", var password: String? = "")