+++ CHANGELOG.md
@@ -7,6 +7,6 @@
 ### Added
 
 * welcome screen
-* basuc connector view triggered by the Coder's welcome view.
+* basic connector view triggered by the Coder's welcome view. It asks the user a Coder hostname, port, email and password.
 * back button to return to the main welcome view
-* basic Coder http client
+* basic Coder http client which authenticates and retrieves a session token